+++ CHANGELOG.rst
@@ -1,12 +1,21 @@
 Changelog
 =========
 
+0.5.0 (UNRELEASED)
+------------------
+
+*   Add RegexFilter to suppress logging of messages containing the match of a
+    a given regular expression.
+*   The `host` field will now always contain the fully qualified domain name.
+*   Remove hardcoded `version` field.
+*   Remove support for Python 2.6.
+
 0.4.1 (2017-05-27)
------
-*   Add NamedCeeLogger 
+------------------
+*   Add NamedCeeLogger
 
 0.3.4 (2016-12-23)
------
+------------------
 *   Add JsonFormatter
 
 0.3.3
@@ -15,7 +24,7 @@ Changelog
 *   Custom log record fields with leading underscore are now contained in
     the log message. In case the (otherwise) same custom field exists both
     with and without leading underscore, the one with leading underscore has
-    precedence. 
+    precedence.
 
 0.3.2 (2015-03-18)
 ------------------